shuffle some code, add some comments

+++ esynth_consts.h
@@ -7,7 +7,7 @@ const int max_trys_per_pixel     = 40000;
 const int max_comp_patch_radius      = 3;
 const int max_transfer_patch_radius  = 2;
 
-const int important_count        = 6;
+const size_t important_count        = 6;
 
 #endif // ESYNTH_CONSTS_H
